% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunction.R
\name{plotMDS}
\alias{plotMDS}
\title{Plots the results of a multidimensional scaling analysis for a DESeqTransform object.}
\usage{
plotMDS(rld)
}
\arguments{
\item{rld}{A \code{\link{DESeqTransform}} object.}
}
\value{
A \code{\link{ggplot}} object.
}
\description{
This function inputs a \code{\link{DESeqTransform}} object, runs an MDS analysis
and plots the result.
}
\examples{
library(DESeq2)
dds <- makeExampleDESeqDataSet(betaSD=1)
rld <- rlog(dds)
plotMDS(rld)
}
\author{
Alejandro Reyes
}
\keyword{MDS}
